consolidate code samples
